OCR master/requirements.txt
google-cloud-vision==3.7.0
pdf2image==1.17.0
pillow==10.3.0
python-dotenv==1.0.0
requests==2.32.5
